refine makefiles so apps only linke to dsn.core.so
