+++ ext/gtest/CMakeLists.txt
@@ -8,7 +8,7 @@ ExternalProject_Add(googletest
    #   CMAKE_ARGS -DCMAKE_ARCHIVE_OUTPUT_DIRECTORY_DEBUG:PATH=DebugLibs
    #            -DCMAKE_ARCHIVE_OUTPUT_DIRECTORY_RELEASE:PATH=ReleaseLibs
    #            -DCMAKE_CXX_FLAGS=${MSVC_COMPILER_DEFS}
-    CMAKE_ARGS -Dgtest_force_shared_crt=OFF
+    CMAKE_ARGS "${CMAKE_ARGS};-Dgtest_force_shared_crt=OFF;-DCMAKE_CXX_FLAGS=-fPIC ${CMAKE_CXX_FLAGS}"
     BINARY_DIR "${PROJECT_BINARY_DIR}/lib"
 # Disable install step
     INSTALL_COMMAND ""
