+++ src/core/tests/CMakeLists.txt
@@ -16,11 +16,12 @@ set(MY_PROJ_INC_PATH
 	)
 	
 set(MY_PROJ_LIBS gtest
-                 dsn.corelib
-				 dsn.dev.cpp.core.use
-                 dsn.tools.common
+                 dsn.tools.nfs
+                 dsn.dev.cpp.core.use
                  dsn.tools.simulator
-                 dsn.tools.nfs)
+                 dsn.tools.common
+                 dsn.corelib
+  )
 if(${CMAKE_SYSTEM_NAME} STREQUAL "Linux")
 	set(MY_PROJ_LIBS ${MY_PROJ_LIBS} dl)
 endif()
@@ -31,3 +32,5 @@ set(MY_PROJ_LIB_PATH "")
 set(MY_BINPLACES "")
 
 dsn_add_executable()
+
+
